more cache, storage and database refactoring

+++ backend/database/likes.go
@@ -6,6 +6,7 @@ import (
 	log "github.com/Sirupsen/logrus"
 )
 
+// Likes Data Access Object - provides operations on Likes database table
 type LikesDAO interface {
 	LikeTweet(tweetID, userID int64) error
 	UnlikeTweet(tweetID, userID int64) error
@@ -39,10 +40,7 @@ func (db *likesDB) LikeTweet(tweetID, userID int64) error {
 }
 
 func (db *likesDB) UnlikeTweet(tweetID, userID int64) error {
-	_, err := db.Exec(`
-		DELETE FROM likes
-		WHERE tweet_id=$1 AND user_id=$2`,
-		tweetID, userID)
+	_, err := db.Exec(`DELETE FROM likes WHERE tweet_id=$1 AND user_id=$2`, tweetID, userID)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"tweetID": tweetID,
@@ -57,12 +55,7 @@ func (db *likesDB) UnlikeTweet(tweetID, userID int64) error {
 func (db *likesDB) LikeCount(tweetID int64) (int64, error) {
 	var likeCount int64
 
-	err := db.QueryRow(`
-		SELECT COUNT(*)
-		FROM likes
-		WHERE tweet_id = $1`,
-		tweetID).
-		Scan(&likeCount)
+	err := db.QueryRow(`SELECT COUNT(*) FROM likes WHERE tweet_id = $1`, tweetID).Scan(&likeCount)
 	if err != nil {
 		log.WithField("tweetID", tweetID).WithError(err).Error("LikeCount query error.")
 		return 0, err
@@ -74,13 +67,8 @@ func (db *likesDB) LikeCount(tweetID int64) (int64, error) {
 func (db *likesDB) IsLiked(tweetID, userID int64) (bool, error) {
 	var isLiked bool
 
-	err := db.QueryRow(`
-		SELECT exists
-			(SELECT true
-			FROM likes
-			WHERE tweet_id = $1 AND user_id = $2)`,
-		tweetID, userID).
-		Scan(&isLiked)
+	err := db.QueryRow(`SELECT exists (SELECT TRUE FROM likes WHERE tweet_id = $1 AND user_id = $2)`,
+		tweetID, userID).Scan(&isLiked)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"tweetID": tweetID,
